+++ serpentTools/parsers/depletion.py
@@ -33,7 +33,7 @@ METADATA_KEYS = {'ZAI', 'NAMES', 'BU', 'DAYS'}
 class DepPlotMixin(object):
 
     @magicPlotDocDecorator
-    def plot(self, xUnits, yUnits, timePoints=None, names=None, zai=None,
+    def plot(self, xUnits, yUnits=None, timePoints=None, names=None, zai=None,
              materials=None, ax=None, legend=None, logx=False, logy=False,
              loglog=False, labelFmt=None, xlabel=None, ylabel=None, ncol=1,
              **kwargs):
@@ -43,12 +43,18 @@ class DepPlotMixin(object):
         Parameters
         ----------
         xUnits: str
+            If ``xUnits`` is given and ``yUnits`` is ``None``, then
+            the plotted data will be ``xUnits`` against ``'days'``
             name of x value to obtain, e.g. ``'days'``, ``'burnup'``
         yUnits: str
             name of y value to return, e.g. ``'adens'``, ``'ingTox'``
         timePoints: list or None
             If given, select the time points according to those
             specified here. Otherwise, select all points
+
+            .. deprecated:: 0.7.0
+               Will plot against all time points
+
         names: str or list or None
             If given, plot  values corresponding to these isotope
             names. Otherwise, plot values for all isotopes.
@@ -90,6 +96,10 @@ class DepPlotMixin(object):
         SerpentToolsException
             If the materials dictionary does not contain any items
         """
+        if yUnits is None:
+            yUnits = xUnits
+            xUnits = 'days'
+
         if not self.materials:
             raise SerpentToolsException("Material dictionary is empty")